Refactor ImportFlowPage and routes
