+++ src/components/ImportFlowPage/ImportFlowPage.tsx
@@ -5,43 +5,8 @@ import { useTranslation } from 'react-i18next';
 import { AltRoute, ArrowBack } from '@mui/icons-material';
 import { useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import { PageRoutes } from '../../routes/routes';
+import { importFlowConfig } from '../../routes/routes';
 import TitleComponent from '../TitleComponent/TitleComponent';
-
-interface ImportFlowDetail {
-  title: string,
-  fileExtensionsAllowed: string[],
-  backRoute: string,
-  successRoute: string,
-  requiredFieldDescription?: string,
-  flowTypes?: string[]
-}
-
-type ImportFlowDetails = Record<string, ImportFlowDetail>
-
-
-const importFlowConfig: ImportFlowDetails = {
-  'telematic-receipt': {
-    title: 'commons.routes.TELEMATIC_RECEIPT_IMPORT_FLOW',
-    fileExtensionsAllowed: ['zip'],
-    backRoute: PageRoutes.TELEMATIC_RECEIPT_IMPORT_OVERVIEW,
-    successRoute: PageRoutes.TELEMATIC_RECEIPT_IMPORT_FLOW_THANK_YOU_PAGE
-  },
-  'reporting': {
-    title: 'commons.routes.REPORTING_IMPORT_FLOW',
-    fileExtensionsAllowed: ['zip'],
-    backRoute: PageRoutes.REPORTING_IMPORT_OVERVIEW,
-    successRoute: PageRoutes.REPORTING_IMPORT_FLOW_THANK_YOU_PAGE
-  },
-  'treasury': {
-    title: 'commons.routes.TREASURY_IMPORT_FLOW',
-    fileExtensionsAllowed: ['zip'],
-    backRoute: PageRoutes.TREASURY,
-    successRoute: PageRoutes.REPORTING_IMPORT_FLOW_THANK_YOU_PAGE,
-    requiredFieldDescription: 'commons.requiredFieldDescription',
-    flowTypes: ['Giornale di Cassa XLS', 'Giornale di Cassa CSV', 'Giornale di Cassa OPI', 'Estrato conto poste']
-  }
-};
 
 const ImportFlow = () => {
